requirements.txt
torch==1.13.1
transformers==4.30.2
pillow==9.5.0
pandas==1.1.5
